+++ test/methods_test.dart
@@ -10,10 +10,10 @@ void main() {
       () async => expect(
         fn(),
         completion(
-          isA<MySuccess<String>>()
+          isA<MyHttpSuccess<String>>()
               .having((s) => s.result, 'result', method)
               .having(
-                (s) => s.response?.requestOptions.method,
+                (s) => s.response.requestOptions.method,
                 'response.requestOptions.method',
                 method,
               ),